+++ tests/SwaggerServiceTest.php
@@ -10,10 +10,14 @@ use RonasIT\AutoDoc\Exceptions\LegacyConfigException;
 use RonasIT\AutoDoc\Exceptions\SwaggerDriverClassNotFoundException;
 use RonasIT\AutoDoc\Exceptions\UnsupportedDocumentationViewerException;
 use RonasIT\AutoDoc\Exceptions\WrongSecurityConfigException;
+use RonasIT\AutoDoc\Http\Middleware\AutoDocMiddleware;
 use RonasIT\AutoDoc\Services\SwaggerService;
 use RonasIT\AutoDoc\Tests\Support\Mock\TestContract;
 use RonasIT\AutoDoc\Tests\Support\Mock\TestNotificationSetting;
 use RonasIT\AutoDoc\Tests\Support\Mock\TestRequest;
+use RonasIT\AutoDoc\Tests\Support\Models\User;
+use RonasIT\AutoDoc\Tests\Support\Resources\UserResource;
+use RonasIT\AutoDoc\Tests\Support\Resources\UsersCollectionResource;
 use RonasIT\AutoDoc\Tests\Support\Traits\SwaggerServiceMockTrait;
 use RonasIT\AutoDoc\Tests\Support\Traits\SwaggerServiceTestingTrait;
 use stdClass;
@@ -982,5 +986,66 @@ class SwaggerServiceTest extends TestCase
         ]);
 
         app(SwaggerService::class)->addData($request, $response);
+    }
+
+    public function testHandleResponseWithResource()
+    {
+        $this->mockDriverGetEmptyAndSaveProcessTmpData($this->getJsonFixture('response_with_resource'));
+
+        $request = $this->generateRequest(
+            type: 'get',
+            uri: '/user',
+            controllerMethod: 'user',
+        );
+
+        $resource = UserResource::make(User::factory()->make());
+
+        app(SwaggerService::class)->addData($request, $resource->toResponse($request));
+    }
+
+    public function testHandleResponseWithResourceCollection()
+    {
+        $this->mockDriverGetEmptyAndSaveProcessTmpData($this->getJsonFixture('response_with_resource_collection'));
+
+        $request = $this->generateRequest(
+            type: 'get',
+            uri: '/users',
+            controllerMethod: 'users',
+        );
+
+        $resource = UsersCollectionResource::make(collect([
+            User::factory()->make(),
+            User::factory()->make(),
+        ]));
+
+        app(SwaggerService::class)->addData($request, $resource->toResponse($request));
+    }
+
+    public function testHandleResponseNotResource()
+    {
+        $this->mockDriverGetEmptyAndSaveProcessTmpData($this->getJsonFixture('response_not_resource'));
+
+        $request = $this->generateRequest(
+            type: 'delete',
+            uri: '/users',
+            controllerMethod: 'deleteProfile',
+        );
+
+        app(SwaggerService::class)->addData($request, response()->noContent());
+    }
+
+    public function testHandleResponseAliasToResource()
+    {
+        $this->mockDriverGetEmptyAndSaveProcessTmpData($this->getJsonFixture('response_with_resource'));
+
+        $request = $this->generateRequest(
+            type: 'get',
+            uri: '/user',
+            controllerMethod: 'userAliasResource',
+        );
+
+        $resource = UserResource::make(User::factory()->make());
+
+        app(SwaggerService::class)->addData($request, $resource->toResponse($request));
     }
 }
